Simplify code and refactor
